move pinyin to gpy.go and refactor code
