+++ pinyin.go
@@ -4,22 +4,7 @@ import (
 	// "fmt"
 	"regexp"
 	"strings"
-	"unicode"
-
-	"github.com/go-ego/gse"
 )
-
-// Meta
-const (
-	// Version get the gpy version
-	Version = "v0.20.0.72"
-	// License   = "MIT"
-)
-
-// GetVersion get the version
-func GetVersion() string {
-	return Version
-}
 
 // 拼音风格(推荐)
 const (
@@ -35,64 +20,11 @@ const (
 	FinalsTone3 = 9 // 韵母风格3，带声调，声调在各个拼音之后，用数字 [1-4] 进行表示。如： ong1 uo2
 )
 
-// 拼音风格(兼容之前的版本)
-// const (
-// 	NORMAL       = Normal
-// 	TONE         = Tone
-// 	TONE2        = Tone2
-// 	INITIALS     = Initials
-// 	FIRST_LETTER = FirstLetter
-// 	FINALS       = Finals
-// 	FINALS_TONE  = FinalsTone
-// 	FINALS_TONE2 = FinalsTone2
-// )
-
-var (
-	// 声母表
-	initialArray = strings.Split(
-		"b,p,m,f,d,t,n,l,g,k,h,j,q,x,r,zh,ch,sh,z,c,s",
-		",",
-	)
-
-	// 所有带声调的字符
-	rePhoneticSymbolSource = func(m map[string]string) string {
-		s := ""
-		for k := range m {
-			s = s + k
-		}
-		return s
-	}(phoneticSymbol)
-)
-
-var (
-	// 匹配带声调字符的正则表达式
-	rePhoneticSymbol = regexp.MustCompile("[" + rePhoneticSymbolSource + "]")
-
-	// 匹配使用数字标识声调的字符的正则表达式
-	reTone2 = regexp.MustCompile("([aeoiuvnm])([1-4])$")
-
-	// 匹配 Tone2 中标识韵母声调的正则表达式
-	reTone3 = regexp.MustCompile("^([a-z]+)([1-4])([a-z]*)$")
-)
-
-// Args 配置信息
-type Args struct {
-	Style     int    // 拼音风格（默认： Normal)
-	Heteronym bool   // 是否启用多音字模式（默认：禁用）
-	Separator string // Slug 中使用的分隔符（默认：-)
-
-	// 处理没有拼音的字符（默认忽略没有拼音的字符）
-	// 函数返回的 slice 的长度为0 则表示忽略这个字符
-	Fallback func(r rune, a Args) []string
-}
-
 var (
 	// Style 默认配置：风格
 	Style = Normal
-
 	// Heteronym 默认配置：是否启用多音字模式
 	Heteronym = false
-
 	// Separator 默认配置： `Slug` 中 Join 所用的分隔符
 	Separator = "-"
 
@@ -108,9 +40,42 @@ var (
 		"ù": "ǜ",
 	}
 
+	// 声母表
+	initialArray = strings.Split(
+		"b,p,m,f,d,t,n,l,g,k,h,j,q,x,r,zh,ch,sh,z,c,s",
+		",",
+	)
+
+	// 所有带声调的字符
+	rePhoneticSymbolSource = func(m map[string]string) string {
+		s := ""
+		for k := range m {
+			s = s + k
+		}
+		return s
+	}(phoneticSymbol)
+
+	// 匹配带声调字符的正则表达式
+	rePhoneticSymbol = regexp.MustCompile("[" + rePhoneticSymbolSource + "]")
+	// 匹配使用数字标识声调的字符的正则表达式
+	reTone2 = regexp.MustCompile("([aeoiuvnm])([1-4])$")
+	// 匹配 Tone2 中标识韵母声调的正则表达式
+	reTone3 = regexp.MustCompile("^([a-z]+)([1-4])([a-z]*)$")
+
 	reFinalExceptions  = regexp.MustCompile("^(j|q|x)(ū|ú|ǔ|ù)$")
 	reFinal2Exceptions = regexp.MustCompile("^(j|q|x)u(\\d?)$")
 )
+
+// Args 配置信息
+type Args struct {
+	Style     int    // 拼音风格（默认： Normal)
+	Heteronym bool   // 是否启用多音字模式（默认：禁用）
+	Separator string // Slug 中使用的分隔符（默认：-)
+
+	// 处理没有拼音的字符（默认忽略没有拼音的字符）
+	// 函数返回的 slice 的长度为0 则表示忽略这个字符
+	Fallback func(r rune, a Args) []string
+}
 
 // NewArgs 返回包含默认配置的 `Args`
 func NewArgs() Args {
@@ -219,148 +184,4 @@ func applyStyle(p []string, a Args) []string {
 		newP = append(newP, ToFixed(v, a))
 	}
 	return newP
-}
-
-// SinglePinyin 把单个 `rune` 类型的汉字转换为拼音.
-func SinglePinyin(r rune, a Args) []string {
-	if a.Fallback == nil {
-		a.Fallback = Fallback
-	}
-	value, ok := PinyinDict[int(r)]
-	pys := []string{}
-	if ok {
-		pys = strings.Split(value, ",")
-	} else {
-		pys = a.Fallback(r, a)
-	}
-	if len(pys) > 0 {
-		if !a.Heteronym {
-			pys = pys[:1]
-		}
-
-		return applyStyle(pys, a)
-	}
-	return pys
-}
-
-// IsChineseChar to determine whether the Chinese string
-// 判断是否为中文字符串
-func IsChineseChar(str string) bool {
-	for _, r := range str {
-		if unicode.Is(unicode.Scripts["Han"], r) {
-			return true
-		}
-	}
-	return false
-}
-
-func args(arg ...Args) Args {
-	a := NewArgs()
-	if len(arg) > 0 {
-		a = arg[0]
-	}
-
-	return a
-}
-
-// HanPinyin 汉字转拼音，支持多音字模式.
-func HanPinyin(s string, arg ...Args) [][]string {
-	a := args(arg...)
-
-	pys := [][]string{}
-	for _, r := range s {
-		py := SinglePinyin(r, a)
-		if len(py) > 0 {
-			pys = append(pys, py)
-		}
-	}
-	return pys
-}
-
-// Pinyin 汉字转拼音，支持多音字模式、拼音与英文等字母混合.
-func Pinyin(s string, arg ...Args) [][]string {
-	a := args(arg...)
-
-	pys := [][]string{}
-	sw := gse.SplitTextToWords([]byte(s))
-	for i := 0; i < len(sw); i++ {
-		s1 := string([]byte(sw[i]))
-		r := []rune(s1)
-		if len(r) <= 1 && unicode.Is(unicode.Han, r[0]) {
-			py := SinglePinyin(r[0], a)
-			if len(py) > 0 {
-				pys = append(pys, py)
-			}
-		} else {
-			var pya []string
-			pya = append(pya, s1)
-			pys = append(pys, pya)
-		}
-	}
-
-	return pys
-}
-
-// ToString trans pinyin [][]string to string
-func ToString(p [][]string) (s string) {
-	i := 0
-	for _, p1 := range p {
-		r := []rune(p1[0])[0]
-		i++
-
-		if unicode.IsLetter(r) && i > 1 {
-			s += " " + p1[0]
-		} else {
-			if i > 1 || unicode.IsSpace(r) {
-				i = 0
-			}
-			s += p1[0]
-		}
-	}
-
-	return
-}
-
-// Py return to string pinyin
-func Py(s string, a ...Args) string {
-	p := Pinyin(s, a...)
-	return ToString(p)
-}
-
-// LazyPinyin 汉字转拼音，与 `Pinyin` 的区别是：
-// 返回值类型不同，并且不支持多音字模式，每个汉字只取第一个音.
-func LazyPinyin(s string, arg ...Args) []string {
-	a := args(arg...)
-
-	a.Heteronym = false
-	pys := []string{}
-	for _, v := range HanPinyin(s, a) {
-		pys = append(pys, v[0])
-	}
-	return pys
-}
-
-// Slug join `LazyPinyin` 的返回值.
-// 建议改用 https://github.com/mozillazg/go-slugify
-func Slug(s string, a Args) string {
-	separator := a.Separator
-	return strings.Join(LazyPinyin(s, a), separator)
-}
-
-// Convert 跟 Pinyin 的唯一区别就是 a 参数可以是 nil
-func Convert(s string, a *Args) [][]string {
-	if a == nil {
-		args := NewArgs()
-		a = &args
-	}
-	return Pinyin(s, *a)
-}
-
-// LazyConvert 跟 LazyPinyin 的唯一区别就是 a 参数可以是 nil
-func LazyConvert(s string, a *Args) []string {
-	if a == nil {
-		args := NewArgs()
-		a = &args
-	}
-	return LazyPinyin(s, *a)
 }
